got the edit texture to render correctly from the PAL surface, will be working on changing it to edit the texture directly next

+++ msk2bmpGUI_C_C++_imgui_sdl2_tinyfiledialogs/Palette_Cycle.h
@@ -6,4 +6,4 @@
 //color cycling
 bool load_palette_to_array(float* data);
 void update_palette_array(float* palette, double CurrentTime, bool* Palette_Update);
-void Color_Cycle(float* PaletteColors, uint16_t* g_dwCurrent, int pal_index, uint8_t * cycle_colors, int cycle_count);
+void Color_Cycle(float* PaletteColors, int* g_dwCurrent, int pal_index, uint8_t * cycle_colors, int cycle_count);
